% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdynmc_exploratory.R
\name{data.info}
\alias{data.info}
\title{Investigate if panel data set is balanced or unbalanced.}
\usage{
<<<<<<< HEAD
data.info(object, i.name = NULL, t.name = NULL, ...)
}
\arguments{
\item{object}{An object of class `data.frame`.}
=======
data.info(object, i.name, t.name, ...)
}
\arguments{
\item{object}{An object of class `pdynmc`.}
>>>>>>> 1d2dbc567a34f8efe2d5f7834b7ebd3c2145ff15

\item{i.name}{Column name of cross-section identifier.}

\item{t.name}{Column name of time-series identifier.}

\item{...}{further arguments.}
}
\value{
<<<<<<< HEAD
Returns information if panel data set contained
   in an object of class `data.frame` is a balanced or
   unbalanced panel data set.
}
\description{
\code{data.info} Investigate if a panel data set contained
   in a `data.frame` is balanced or unbalanced.
=======
Returns information if object of class `data.frame`
   is a balanced or unbalanced panel data set.
}
\description{
\code{data.info} Plot fitted values against residuals for an
   object of class `pdynmc`.
>>>>>>> 1d2dbc567a34f8efe2d5f7834b7ebd3c2145ff15
}
\examples{
## Load data from plm package
if(!requireNamespace("plm", quietly = TRUE)){
 stop("Dataset from package \"plm\" needed for this example. Please install the package.", call. = FALSE)
} else{
 data(EmplUK, package = "plm")
 dat <- EmplUK
 dat[,c(4:7)] <- log(dat[,c(4:7)])
 dat <- dat[c(1:140), ]

## Code example
 data.info(dat, i.name = "firm", t.name = "year")

 data.info(dat[dat$year \%in\% 1979:1981, ], i.name = "firm", t.name = "year")
}


}
\seealso{
\code{\link{pdynmc}} for fitting a linear dynamic panel data model.
}
<<<<<<< HEAD
\author{
Markus Fritsch, Joachim Schnurbus
}
=======
>>>>>>> 1d2dbc567a34f8efe2d5f7834b7ebd3c2145ff15
