% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdynmc-package.R
\docType{package}
\name{pdynmc-package}
\alias{pdynmc}
\alias{pdynmc-package}
\title{pdynmc: Moment Condition Based Estimation of Linear Dynamic Panel Data Models}
\description{
Linear dynamic panel data modeling based on linear and nonlinear moment conditions as proposed by Holtz-Eakin, Newey, and Rosen (1988) \doi{10.2307/1913103}, Ahn and Schmidt (1995) \doi{10.1016/0304-4076(94)01641-C}, and Arellano and Bover (1995) \doi{10.1016/0304-4076(94)01642-D}. Estimation of the model parameters relies on the Generalized Method of Moments (GMM), numerical optimization (when nonlinear moment conditions are employed) and the computation of closed form solutions (when estimation is based on linear moment conditions). One-step, two-step and iterated estimation is available. For inference and specification testing, Windmeijer (2005) \doi{10.1016/j.jeconom.2004.02.005} and doubly corrected standard errors (Hwang, Kang, Lee, 2021 \doi{10.1016/j.jeconom.2020.09.010}) are available. Additionally, serial correlation tests, tests for overidentification, and Wald tests are provided. Functions for visualizing panel data structures and modeling results obtained from GMM estimation are also available. The plot methods include functions to plot unbalanced panel structure, coefficient ranges and coefficient paths across GMM iterations (the latter is implemented according to the plot shown in Hansen and Lee, 2021 \doi{10.3982/ECTA16274}). For a more detailed description of the functionality, please see Fritsch, Pua, Schnurbus (2021) \doi{10.32614/RJ-2021-035}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/markusfritsch/pdynmc}
  \item Report bugs at \url{https://github.com/markusfritsch/pdynmc/issues}
}

}
\author{
\strong{Maintainer}: Markus Fritsch \email{Markus.Fritsch@uni-Passau.de}

Authors:
\itemize{
  \item Joachim Schnurbus \email{Joachim.Schnurbus@uni-Passau.de}
  \item Andrew Adrian Yu Pua \email{andrewypua@gmail.com}
}

}
\keyword{internal}
