% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-pdynmc.R
\docType{package}
\name{pdynmc}
\alias{pdynmc}
\title{pdynmc: A package for moment conditions based estimation of linear dynamic
   panel data models}
\description{
The pdynmc package provides three categories of functions that are available
   to the user: A function for model fitting, functions for specification
   testing, and functions that extract and summarize particular information
   from fitted model objects.
}
