% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdynmc_fitMethods.R
\name{plot.coef.pdynmc}
\alias{plot.coef.pdynmc}
\title{Plot of coefficient estimates and corresponding ranges.}
\usage{
\method{plot}{coef.pdynmc}(
  object,
  include.dum = FALSE,
  include.fur.con = FALSE,
  col.coefRange = 1,
  col.coefInitial = "darkgrey",
  col.coefEst = "royalblue",
  boxplot.coef = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class `pdynmc`. The function requires
twostep or iterative GMM estimates.}

\item{include.dum}{Include estimates of parameters corresponding to time
dummies (defaults to 'FALSE').}

\item{include.fur.con}{Include estimates of parameters corresponding to
further controls (defaults to 'FALSE').}

\item{col.coefRange}{Specify color for plotting range of coefficient
estimates (defaults to 'black').}

\item{col.coefInitial}{Specify color for plotting initial coefficient
estimate (defaults to 'darkgrey').}

\item{col.coefEst}{Specify color for plotting coefficient estimate
(defaults to 'royalblue').}

\item{boxplot.coef}{Wether to draw boxplots for coefficient estimates
(defaults to 'FALSE'); requires iterative GMM with at least 10
iterations.}

\item{...}{further arguments.}
}
\value{
Plot coefficient estimates and coefficient estimate ranges
   for object of class `pdynmc` (requires twostep or iterative GMM
   estimates).
}
\description{
\code{plot.coef.pdynmc} Plot coefficient estimates and corresponding
   coefficient estimate ranges for objects of class `pdynmc`.
}
\examples{
## Load data from plm package
if(!requireNamespace("plm", quietly = TRUE)){
 stop("Dataset from package \"plm\" needed for this example. Please install the package.", call. = FALSE)
} else{
 data(EmplUK, package = "plm")
 dat <- EmplUK
 dat[,c(4:7)] <- log(dat[,c(4:7)])
 dat <- dat[c(1:140), ]

## Code example
 m1 <- pdynmc(dat = dat, varname.i = "firm", varname.t = "year",
    use.mc.diff = TRUE, use.mc.lev = FALSE, use.mc.nonlin = FALSE,
    include.y = TRUE, varname.y = "emp", lagTerms.y = 2,
    fur.con = TRUE, fur.con.diff = TRUE, fur.con.lev = FALSE,
    varname.reg.fur = c("wage", "capital", "output"), lagTerms.reg.fur = c(1,2,2),
    include.dum = TRUE, dum.diff = TRUE, dum.lev = FALSE, varname.dum = "year",
    w.mat = "iid.err", std.err = "corrected", estimation = "twostep",
    opt.meth = "none")
 plot.coef(m1)
}

\donttest{
## Load data from plm package
if(!requireNamespace("plm", quietly = TRUE)){
 stop("Dataset from package \"plm\" needed for this example. Please install the package.", call. = FALSE)
} else{
 data(EmplUK, package = "plm")
 dat <- EmplUK
 dat[,c(4:7)] <- log(dat[,c(4:7)])

## Further code example
 m1 <- pdynmc(dat = dat, varname.i = "firm", varname.t = "year",
    use.mc.diff = TRUE, use.mc.lev = FALSE, use.mc.nonlin = FALSE,
    include.y = TRUE, varname.y = "emp", lagTerms.y = 2,
    fur.con = TRUE, fur.con.diff = TRUE, fur.con.lev = FALSE,
    varname.reg.fur = c("wage", "capital", "output"), lagTerms.reg.fur = c(1,2,2),
    include.dum = TRUE, dum.diff = TRUE, dum.lev = FALSE, varname.dum = "year",
    w.mat = "iid.err", std.err = "corrected", estimation = "iterative",
    opt.meth = "none")
 plot.coef(m1)
}
}


}
\seealso{
\code{\link{pdynmc}} for fitting a linear dynamic panel data model.
}
