% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdynmc_fitMethods.R
\name{plot.pdynmc}
\alias{plot.pdynmc}
\title{Plot coefficient estimates and ranges of coefficient estimates.}
\usage{
\method{plot}{pdynmc}(
  object,
  include.dum = FALSE,
  include.fur.con = FALSE,
  col.coefRange = 1,
  col.coefEst = "royalblue",
  boxplot.coef = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class `pdynmc`.}

\item{include.dum}{Include estimates of parameters corresponding to time
dummies (defaults to 'FALSE')}

\item{include.fur.con}{Include estimates of parameters corresponding to
further controls (defaults to 'FALSE')}

\item{col.coefRange}{Specify color for plotting range of coefficient
estimates (defaults to 'black')}

\item{boxplot.coef}{Wether to draw boxplots for coefficient estimates
  (defaults to 'FALSE'); requires iterative GMM with at least 10
  iterations
An integer denoting the iteration step for which input parameters
  are extracted (defaults to last iteration step used for obtaining parameter
  estimates).}

\item{...}{further arguments.}

\item{coel.coefEst}{Specify color for plotting coefficient estimate
(defaults to 'royalblue')}
}
\value{
Plot coefficients and coefficient range of object of class
   `pdynmc`.
}
\description{
\code{plot.pdynmc} Plot coefficient estimates and ranges of coefficient
   estimates of object of class `pdynmc` (requires at twostep or
   iterative GMM estimates).
}
\examples{
## Load data from plm package
if(!requireNamespace("plm", quietly = TRUE)){
 stop("Dataset from package \"plm\" needed for this example. Please install the package.", call. = FALSE)
} else{
 data(EmplUK, package = "plm")
 dat <- EmplUK
 dat[,c(4:7)] <- log(dat[,c(4:7)])
 dat <- dat[c(1:140), ]

## Code example
 m1 <- pdynmc(dat = dat, varname.i = "firm", varname.t = "year",
    use.mc.diff = TRUE, use.mc.lev = FALSE, use.mc.nonlin = FALSE,
    include.y = TRUE, varname.y = "emp", lagTerms.y = 2,
    fur.con = TRUE, fur.con.diff = TRUE, fur.con.lev = FALSE,
    varname.reg.fur = c("wage", "capital", "output"), lagTerms.reg.fur = c(1,2,2),
    include.dum = TRUE, dum.diff = TRUE, dum.lev = FALSE, varname.dum = "year",
    w.mat = "iid.err", std.err = "corrected", estimation = "twostep",
    opt.meth = "none")
 plot.pdynmc(m1)
}

\donttest{
## Load data from plm package
if(!requireNamespace("plm", quietly = TRUE)){
 stop("Dataset from package \"plm\" needed for this example. Please install the package.", call. = FALSE)
} else{
 data(EmplUK, package = "plm")
 dat <- EmplUK
 dat[,c(4:7)] <- log(dat[,c(4:7)])

## Further code example
 m1 <- pdynmc(dat = dat, varname.i = "firm", varname.t = "year",
    use.mc.diff = TRUE, use.mc.lev = FALSE, use.mc.nonlin = FALSE,
    include.y = TRUE, varname.y = "emp", lagTerms.y = 2,
    fur.con = TRUE, fur.con.diff = TRUE, fur.con.lev = FALSE,
    varname.reg.fur = c("wage", "capital", "output"), lagTerms.reg.fur = c(1,2,2),
    include.dum = TRUE, dum.diff = TRUE, dum.lev = FALSE, varname.dum = "year",
    w.mat = "iid.err", std.err = "corrected", estimation = "iterative",
    opt.meth = "none")
 plot.pdynmc(m1)
}
}


}
\seealso{
\code{\link{pdynmc}} for fitting a linear dynamic panel data model.
}
