% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdynmc_fitMethods.R
\name{print.pdynmc}
\alias{print.pdynmc}
\title{Print objects of class `pdynmc`.}
\usage{
\method{print}{pdynmc}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class `pdynmc`.}

\item{digits}{An integer indicating the maximum number of digits to display in the object.}

\item{...}{further arguments.}
}
\value{
Print objects of class `pdynmc`.
}
\description{
\code{print.pdynmc} prints objects of class `pdynmc`.
}
\seealso{
\code{\link{pdynmc}} for fitting a linear dynamic panel data model.
}
