% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdynmc_fitMethods.R
\name{print.summary.pdynmc}
\alias{print.summary.pdynmc}
\title{Print summary for objects of class `pdynmc`.}
\usage{
\method{print}{summary.pdynmc}(
  x,
  digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{An object of class `summary.pdynmc`.}

\item{digits}{An integer indicating the maximum number of digits to display in the object.}

\item{signif.stars}{Argument is defined as in \code{\link{options}}.}

\item{...}{further arguments.}
}
\value{
Print information on objects of class `summary.pdynmc`.
}
\description{
\code{print.summary.pdynmc} prints the summary for objects of class
   `pdynmc`.
}
\seealso{
\code{\link{pdynmc}} for fitting a linear dynamic panel data model.
}
