% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdynmc_fit.R
\name{residuals.pdynmc}
\alias{residuals.pdynmc}
\title{Extract residuals.}
\usage{
\method{residuals}{pdynmc}(object, step = object$iter, na.rm = FALSE, ...)
}
\arguments{
\item{object}{An object of class `pdynmc.fit`.}

\item{step}{An integer denoting the iteration step for which fitted values
are extracted (defaults to last iteration step used for obtaining parameter
estimates).}

\item{na.rm}{A logical variable indicating whether missing values should be
removed from the vector of fitted values (defaults to `FALSE`).}

\item{...}{further arguments.}
}
\value{
Extract residuals from object of class `pdynmc.fit`.
}
\description{
\code{residual.pdynmc} extracts residuals of an object of class
   `pdynmc.fit`.
}
\examples{
\donttest{
   data(EmplUK, package = "plm")
   dat <- EmplUK
   m1 <- pdynmc.fit(dat = dat, varname.i = "firm", varname.t = "year",
      use.mc.diff = TRUE, use.mc.lev = FALSE, use.mc.nonlin = FALSE,
      include.y = TRUE, varname.y = "emp", lagTerms.y = 2,
      fur.con = TRUE, fur.con.diff = TRUE, fur.con.lev = FALSE,
      varname.reg.fur = c("wage", "capital", "output"), lagTerms.reg.fur = c(1,2,2),
      include.dum = TRUE, dum.diff = TRUE, dum.lev = FALSE, varname.dum = "year",
      w.mat = "iid.err", std.err = "corrected", estimation = "onestep",
      opt.meth = "none")
   residuals.pdynmc(m1, na.rm = TRUE)
}


}
\seealso{
\code{\link{pdynmc.fit}} for fitting a linear dynamic panel data model.
}
