% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdynmc_exploratory.R
\name{strucPD.plot}
\alias{strucPD.plot}
\title{Plot on structure of panel data set.}
\usage{
strucPD.plot(object, i.name = NULL, t.name = NULL, ...)
}
\arguments{
\item{object}{An object of class `data.frame`.}

\item{i.name}{Column name of cross-section identifier.}

\item{t.name}{Column name of time-series identifier.}

\item{...}{further arguments.}
}
\value{
Returns a plot for a panel data set contained in an
   object of class `data.frame` that visualizes the structure
   of the data. Cross-section dimension is plotted on the
   ordinate, longitudinal dimension on the abscissa. Each
   cross-sectional observation is represented by a bar.
   Breaks in the bars represent missing longitudinal
   observations.
}
\description{
\code{strucPD.plot} Plot on cross-section and longtudinal
   structure of an object of class `data.frame` containing
   a panel data set.
}
\examples{
## Load data from plm package
if(!requireNamespace("plm", quietly = TRUE)){
 stop("Dataset from package \"plm\" needed for this example. Please install the package.", call. = FALSE)
} else{
 data(EmplUK, package = "plm")
 dat <- EmplUK
 dat[,c(4:7)] <- log(dat[,c(4:7)])
 dat <- dat[c(1:140), ]

## Code example
 strucPD.plot(dat, i.name = "firm", t.name = "year")

 strucPD.plot(dat[dat$year \%in\% 1979:1981, ], i.name = "firm", t.name = "year")
}


}
\seealso{
\code{\link{pdynmc}} for fitting a linear dynamic panel data model.
}
\author{
Markus Fritsch, Joachim Schnurbus
}
