% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdynmc_fit.R
\name{summary.pdynmc}
\alias{summary.pdynmc}
\title{Summary for objects of class `pdynmc.fit`.}
\usage{
\method{summary}{pdynmc}(object, ...)
}
\arguments{
\item{object}{An object of class `pdynmc.fit`.}

\item{...}{further arguments.}
}
\value{
Object of class `summary.pdynmc`.
}
\description{
\code{summary.pdynmc} generates the summary for objects of class `pdynmc.fit`.
}
\examples{
\donttest{
data(EmplUK, package = "plm")
dat <- EmplUK
dat[,c(4:7)] <- log(dat[,c(4:7)])

m1 <- pdynmc.fit(dat = dat, varname.i = "firm", varname.t = "year",
   use.mc.diff = TRUE, use.mc.lev = FALSE, use.mc.nonlin = FALSE,
   include.y = TRUE, varname.y = "emp", lagTerms.y = 2,
   fur.con = TRUE, fur.con.diff = TRUE, fur.con.lev = FALSE,
   varname.reg.fur = c("wage", "capital", "output"), lagTerms.reg.fur = c(1,2,2),
   include.dum = TRUE, dum.diff = TRUE, dum.lev = FALSE, varname.dum = "year",
   w.mat = "iid.err", std.err = "corrected", estimation = "onestep",
   opt.meth = "none")
summary.pdynmc(m1)
}


}
\seealso{
\code{\link{pdynmc.fit}} for fitting a linear dynamic panel data model.
}
