% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdynmc_exploratory.R
\name{data.info}
\alias{data.info}
\title{Investigate if panel data set is balanced or unbalanced.}
\usage{
data.info(object, i.name, t.name, ...)
}
\arguments{
\item{object}{An object of class `pdynmc`.}

\item{i.name}{Column name of cross-section identifier.}

\item{t.name}{Column name of time-series identifier.}

\item{...}{further arguments.}
}
\value{
Returns information if object of class `data.frame`
   is a balanced or unbalanced panel data set.
}
\description{
\code{data.info} Plot fitted values against residuals for an
   object of class `pdynmc`.
}
\examples{
## Load data from plm package
if(!requireNamespace("plm", quietly = TRUE)){
 stop("Dataset from package \"plm\" needed for this example. Please install the package.", call. = FALSE)
} else{
 data(EmplUK, package = "plm")
 dat <- EmplUK
 dat[,c(4:7)] <- log(dat[,c(4:7)])
 dat <- dat[c(1:140), ]

## Code example
 data.info(dat, i.name = "firm", t.name = "year")

 data.info(dat[dat$year \%in\% 1979:1981, ], i.name = "firm", t.name = "year")
}


}
\seealso{
\code{\link{pdynmc}} for fitting a linear dynamic panel data model.
}
